Add files via upload
This project based on javafx and oop. Here we creating map in new ".txt" file by reading it as an input. Throughout the whole game food is appearing randomly in the map and we need to collect as much as possible of them, and in the we gain our total points.